chore: refactor error framework

+++ omg-idl-gen/files/test-vectors/include_directive/expected.rs
@@ -3,12 +3,10 @@
 pub mod ModuleA {
     use std::vec::Vec;
 
-    #[allow(dead_code)]
-    #[allow(non_camel_case_types)]
-    pub type dim1 = [i32;2 as usize];
+    #[allow(dead_code, non_camel_case_types)]
+    pub type dim1 = [i32;2_usize];
 
-    #[allow(dead_code)]
-    #[allow(non_camel_case_types)]
+    #[allow(dead_code, non_camel_case_types)]
     pub type seq_long = Vec<i32>;
 
 }
