+++ omg-idl-gen/files/test-vectors/typedef_array_dim_1/expected.rs
@@ -1,4 +1,3 @@
 
-#[allow(dead_code)]
-#[allow(non_camel_case_types)]
-pub type Foo = [u8;2 as usize];
+#[allow(dead_code, non_camel_case_types)]
+pub type Foo = [u8;2_usize];
